% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mod.PCA}
\alias{mod.PCA}
\title{Principal Component Analysis (PCA)}
\usage{
mod.PCA(D, r, ...)
}
\arguments{
\item{D}{A samples by variables data matrix}

\item{r}{The number of PCA components}

\item{...}{Other arguments. Currently ignored}
}
\value{
The PCA parameters of D
}
\description{
Takes in a samples by variables data matrix and gives the PCA parameters.
}
\author{
Opeoluwa F. Oyedele and Sugnet Gardner-Lubbe
}

