% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PLS.binomial.GLM}
\alias{PLS.binomial.GLM}
\title{Partial Least Squares-Generalized Linear Model (PLS-GLM) fitted for Binomial y}
\usage{
PLS.binomial.GLM(X, y, A, PLS_algorithm = NULL, eps = 1e-04, ...)
}
\arguments{
\item{X}{A (NxP) predictor matrix}

\item{y}{A (Nx1) Binomial-distributed response vector}

\item{A}{The number of PLS components}

\item{PLS_algorithm}{The mod.NIPALS algorithm}

\item{eps}{Cut off value for convergence step}

\item{...}{Other arguments. Currently ignored}
}
\value{
The PLS-GLM parameters of D=[X y]
}
\description{
Takes in a set of predictor variables and a set of response variables and gives the PLS-GLM parameters.
}
\author{
Opeoluwa F. Oyedele and Sugnet Gardner-Lubbe
}

